requirements.txt
asgiref==3.6.0
cyrtranslit==1.1.1
Django==4.2.1
django-filter==23.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
